[css-typed-om] Implement shorthands for .get/.getAll/.has.
This patch implements shorthands for stylemap.get/getAll/has. Shorthands
are implemented as a base CSSStyleValue. They are different to other
properties because they cannot be converted to a single CSSValue.
Instead, we need to store the CSSValues of their longhands so that
they can be set again on the same property.

Patches to come:
- styleMap.set/update
- Serialization of these values.

Bug: 816722
Change-Id: Ia2c3f694809c219566b965ebfae40bc38d74f4c5